+++ quiz/views.py
@@ -1,7 +1,5 @@
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import PermissionDenied
-from django.db.models import IntegerField
-from django.db.models import Sum, F, Case, When
 from django.shortcuts import get_object_or_404, redirect
 from django.shortcuts import render
 from django.views import generic
@@ -20,13 +18,18 @@ class IndexView(generic.ListView):
 
 
 class DetailView(generic.DetailView):
+    """
+    Shows all details of a single Question
+    """
     model = Question
     template_name = "quiz/detail.html"
 
-
-class ResultsView(generic.DetailView):
-    model = Question
-    template_name = "quiz/results.html"
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        question = self.object
+        total_questions = Question.objects.filter(related_quiz=question.related_quiz).count()
+        context['total_questions'] = total_questions
+        return context
 
 
 # --------------------------------------- Leaderboard ----------------------------------
@@ -43,28 +46,33 @@ def leaderboard_quiz_view(request, quiz_id):
 
 
 def _get_leaderboard(quiz):
-    total_questions = Question.objects.filter(related_quiz=quiz).count()
+    turns = QuizTurn.objects.filter(quiz=quiz, is_completed=True)
+    user_scores = list(map(lambda turn: _get_user_score(turn), turns))
+    user_scores = sorted(user_scores, key=lambda t: t["total_correct"], reverse=True)
+    distinct_key = "username"
+    seen_keys = set()
+    distinct_data = []
+    for d in user_scores:
+        k = d[distinct_key]
+        if k not in seen_keys:
+            distinct_data.append(d)
+            seen_keys.add(k)
 
-    user_scores = UserAnswer.objects.filter(
-        turn__quiz=quiz,
-        turn__is_completed=True
-    ).values(
-        'user__username'
-    ).annotate(
-        total_correct=Sum(
-            Case(
-                When(selected_choice__is_correct=True, then=1),
-                default=0,
-                output_field=IntegerField()
-            )
-        ),
-    ).annotate(
-        score_percentage=F('total_correct') * 100.0 / total_questions
-    ).order_by('-score_percentage')  # [:10]
     return {
         'quiz': quiz,
-        'user_scores': user_scores
+        'user_scores': distinct_data
     }
+
+
+def _get_user_score(turn):
+    correct_user_answers = (UserAnswer.objects.filter(turn=turn, selected_choice__is_correct=True)
+                            .values('question').distinct().count())
+    total_questions = Question.objects.filter(related_quiz=turn.quiz).count()
+    return {"total_correct": correct_user_answers,
+            "total_possible": total_questions,
+            "score_percentage": correct_user_answers * 100.0 / total_questions,
+            "username": turn.user.username
+            }
 
 
 # --------------------------------------- Quiz ----------------------------------
@@ -89,21 +97,30 @@ def question_view(request, turn_id, question_id):
     if request.user != turn.user:
         raise PermissionDenied
     question = get_object_or_404(Question, id=question_id)
-    choices = question.choice_set.all()
-
+    already_selected_user_answers = list(UserAnswer.objects.filter(
+        user=turn.user,
+        turn=turn,
+        question=question).all())
+    already_selected_choices = list(map(lambda ua: ua.selected_choice, already_selected_user_answers))
     if request.method == 'POST':
-        form = ChoiceForm(request.POST, question=question)
+        form = ChoiceForm(request.POST, question=question, pre_selection=already_selected_choices)
         if form.is_valid():
             selected_choices = form.cleaned_data['choices']
+            # Handle add or update
             for selected_choice in selected_choices:
-                UserAnswer.objects.update_or_create(
+                answer, _ = UserAnswer.objects.update_or_create(
                     user=turn.user,
                     turn=turn,
                     question=question,
-                    defaults={
-                        'selected_choice': selected_choice
-                    }
+                    selected_choice=selected_choice
                 )
+                if selected_choice in already_selected_choices:
+                    already_selected_user_answers.remove(answer)
+
+            # Handle remove
+            for answer in already_selected_user_answers:
+                answer.delete()
+
             next_question = turn.quiz.question_set.filter(id__gt=question.id).first()
             if next_question:
                 return redirect('question', turn_id=turn.id, question_id=next_question.id)
@@ -112,10 +129,11 @@ def question_view(request, turn_id, question_id):
                 turn.save()
                 return redirect('results', turn_id=turn.id)
     else:
-        form = ChoiceForm(question=question)
+        form = ChoiceForm(question=question, pre_selection=already_selected_choices)
 
     previous_question = turn.quiz.question_set.filter(id__lt=question.id).last()
     next_question = turn.quiz.question_set.filter(id__gt=question.id).first()
+    choices = question.choice_set.all()
 
     return render(request, 'quiz/question.html', {
         'quiz': turn,
@@ -137,15 +155,11 @@ def results_view(request, turn_id):
     for question in questions:
         user_answers = UserAnswer.objects.filter(user=user, question=question, turn=turn).all()
         solutions = Choice.objects.filter(question=question, is_correct=True).all()
-        correct_ones = []
-        for answer in user_answers:
-            if answer.selected_choice.is_correct:
-                correct_ones.append(answer.selected_choice)
         results.append({
             'question': question,
             'user_answer': ", ".join(list(
                 map(lambda a: f"'{a.selected_choice.answer_symbol}. {a.selected_choice.choice_text}'", user_answers))),
-            'correct': len(correct_ones) == len(solutions),
+            'correct': len(user_answers.filter(selected_choice__is_correct=True)) == len(solutions),
             'solution': question.solution,
         })
 
@@ -157,18 +171,19 @@ def results_view(request, turn_id):
 
 # --------------------------------------- Problem Reporting ---------------------------------------
 
-def report_problem(request, question_id):
-    question = get_object_or_404(Question, id=question_id)
+def report_problem(request):
     if request.method == 'POST':
-        form = ProblemReportForm(request.POST)
+        form = ProblemReportForm(request.POST, user=request.user if request.user.is_authenticated else None)
+        previous_page = request.POST.get('referer_url', request.META.get('HTTP_REFERER', '/'))
         if form.is_valid():
-            form.save()
-            return redirect('problem_reported')  # Redirect to a success page or the same form
+            problem_report = form.save(commit=False)
+            if request.user.is_authenticated:
+                problem_report.submitted_by = request.user
+            else:
+                problem_report.submitted_by = None
+            problem_report.save()
+            return redirect(previous_page)
     else:
-        form = ProblemReportForm(initial={'question': question})
+        form = ProblemReportForm()
 
-    return render(request, 'quiz/report_problem.html', {'form': form, 'question': question})
-
-
-def problem_reported(request):
-    return render(request, 'quiz/problem_reported.html')
+    return render(request, 'report_problem.html', {'form': form})