+++ changelog.md
@@ -1,5 +1,22 @@
 # JQuiz Changelog
 
+## 0.1
+
+- Added prototypic user account management
+- Fixed result.html shows correct even if answers are incorrect
+- Updated Question detail.html to show solution in hidden area
+- Fixed anonymous user can't report problem
+- Fixed leaderboard showing incorrect numbers
+- Added preselection of `UserAnswers` in form
+- Fixed `UserAnswer` creation when submitting answer to question
+- Removed "Show Results" on last question in list because it would set turn to completed
+- Added `export_quiz "<QUIZ_NAME>"` command
+- Updated to PrismJS `1.29.0`
+- Refactored problem submission to apply everywhere
+- Added fill-in area highlighting
+- Added Menu button to bottom right
+- Fixed results view not showing after quiz completed
+
 ## 0.0.1
 
 Initial proof of concept with minimal number of required features.